[XENFIN-1146] Add HTML texteditor and viewer for comments
