+++ package.json
@@ -40,17 +40,20 @@
     "@material-ui/icons": "^2.0.0",
     "@types/classnames": "^2.2.9",
     "@types/debug": "^0.0.30",
+    "@types/dompurify": "^2.0.1",
     "@types/es6-shim": "^0.31.39",
     "@types/filesize": "^3.2.15",
     "@types/lodash.isequal": "^4.5.5",
     "@types/material-ui": "^0.20.8",
     "@types/node": "^8.10.59",
     "@types/react-dom-factories": "^1.0.2",
+    "@types/react-rte": "^0.16.1",
     "@types/react-table": "^6.8.7",
     "@types/uuid": "^7.0.2",
     "@xenit/finder-utils": "^1.0.2",
     "classnames": "^2.2.6",
     "debug": "^3.2.6",
+    "dompurify": "^2.0.8",
     "filesize": "^3.5.11",
     "flatpickr": "^3.0.7",
     "keycode": "^2.2.0",
@@ -61,6 +64,7 @@
     "react-checkbox-tree": "^1.6.0",
     "react-dom-factories": "^1.0.2",
     "react-flatpickr": "^3.10.0",
+    "react-rte": "^0.16.1",
     "react-sortable-hoc": "^0.8.4",
     "react-table": "^6.11.5",
     "sortablejs": "^1.10.2",
